Move header language helpers to the base component (as in dtq-dev)
Match the dtq-dev structure: getLangCode/getLangCodeIfCzech/translateSlug now
live in the base header component (src/app/header/header.component.ts) and the
themed dspace header just extends it, instead of carrying the helpers itself.

getLangCode() again follows the current/system UI language (read synchronously
from TranslateService, falling back to the configured language before one is
set) rather than being pinned to en/cs. This restores the dtq-dev behaviour.
