+++ src/themes/dspace/app/header/header.component.ts
@@ -1,12 +1,6 @@
-import {
-  Component,
-  inject,
-} from '@angular/core';
+import { Component } from '@angular/core';
 import { RouterLink } from '@angular/router';
-import {
-  TranslateModule,
-  TranslateService,
-} from '@ngx-translate/core';
+import { TranslateModule } from '@ngx-translate/core';
 import { ClarinNavbarTopComponent } from 'src/app/clarin-navbar-top/clarin-navbar-top.component';
 
 import { HeaderComponent as BaseComponent } from '../../../../app/header/header.component';
@@ -30,32 +24,4 @@ import { ImpersonateNavbarComponent } from '../../../../app/shared/impersonate-n
   ],
 })
 export class HeaderComponent extends BaseComponent {
-
-  private readonly translate = inject(TranslateService);
-
-  // UI language for the LINDAT portal links, normalized to the two it supports: 'cs' for Czech, 'en' otherwise.
-  getLangCode(): string {
-    return this.translate.currentLang === 'cs' ? 'cs' : 'en';
-  }
-
-  // Locale segment for portal links: 'cs' in Czech, '' in English.
-  getLangCodeIfCzech(): string {
-    return this.getLangCode() === 'cs' ? 'cs' : '';
-  }
-
-  // Czech portal slug for an English slug; English keeps the original.
-  translateSlug(slug: string): string {
-    if (this.getLangCode() === 'en') {
-      return slug;
-    }
-
-    const translations = {
-      'partners': this.getLangCodeIfCzech() + '/' + 'partneri',
-      'integration': this.getLangCodeIfCzech() + '/' + 'integrace',
-      'partnership': this.getLangCodeIfCzech() + '/' + 'partnerstvi',
-      'services': 'sluzby',
-    };
-
-    return translations[slug] || '';
-  }
 }
